requirements.txt
alembic==0.9.6
aniso8601==9.0.1
bcrypt==3.1.4
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
click==6.7
coverage==4.4.2
dnspython==1.16.0
eventlet==0.31.0
Flask==1.1.4
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.9
Flask-Migrate==2.1.1
flask-restx==0.2.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
flask-talisman==0.8.1
Flask-Testing==0.7.1
google-api-core==1.31.1
google-auth==1.34.0
google-cloud-core==1.7.2
google-cloud-storage==1.41.1
google-crc32c==1.1.2
google-resumable-media==1.3.2
googleapis-common-protos==1.53.0
greenlet==1.1.0
gunicorn==19.7.1
idna==3.2
itsdangerous==0.24
Jinja2==2.11.3
jsonschema==2.6.0
Mako==1.0.7
MarkupSafe==2.0.1
packaging==21.0
protobuf==3.17.3
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.18
PyJWT==1.5.3
pyparsing==2.4.7
python-dateutil==2.6.1
python-editor==1.0.4
pytz==2017.3
requests==2.26.0
rsa==4.7.2
selenium==3.8.1
six==1.16.0
SQLAlchemy==1.4.22
urllib3==1.26.6
Werkzeug==0.15.5
